Add pathFromEvents.ts
