+++ packages/xstate-graph/src/graph.ts
@@ -6,15 +6,12 @@ import type {
   StateFrom,
   EventFrom
 } from 'xstate';
-import { isMachine } from 'xstate/lib/utils';
-import { getAdjacencyMap } from './adjacency';
 import type {
   SerializedEvent,
   SerializedState,
   SimpleBehavior,
   StatePath,
   StatePlan,
-  Steps,
   DirectedGraphEdge,
   DirectedGraphNode,
   TraversalOptions,
@@ -92,7 +89,7 @@ export function createDefaultMachineOptions<TMachine extends AnyStateMachine>(
   };
 }
 
-function createDefaultBehaviorOptions<
+export function createDefaultBehaviorOptions<
   TBehavior extends SimpleBehavior<any, any>
 >(_behavior: TBehavior): TraversalOptions<any, any> {
   return {
@@ -145,88 +142,18 @@ export function toDirectedGraph(
   return graph;
 }
 
-export function getPathFromEvents<
-  TState,
-  TEvent extends EventObject = EventObject
->(
-  behavior: SimpleBehavior<TState, TEvent>,
-  events: TEvent[],
-  options?: TraversalOptions<TState, TEvent>
-): StatePath<TState, TEvent> {
-  const resolvedOptions = resolveTraversalOptions<TState, TEvent>(
-    {
-      getEvents: () => {
-        return events;
-      },
-      ...options
-    },
-    isMachine(behavior)
-      ? createDefaultMachineOptions(behavior)
-      : createDefaultBehaviorOptions(behavior)
-  );
-  const fromState = resolvedOptions.fromState ?? behavior.initialState;
-
-  const { serializeState, serializeEvent } = resolvedOptions;
-
-  const adjacency = getAdjacencyMap(behavior, resolvedOptions);
-
-  const stateMap = new Map<SerializedState, TState>();
-  const path: Steps<TState, TEvent> = [];
-
-  const initialSerializedState = serializeState(
-    fromState,
-    undefined,
-    undefined
-  ) as SerializedState;
-  stateMap.set(initialSerializedState, fromState);
-
-  let stateSerial = initialSerializedState;
-  let state = fromState;
-  for (const event of events) {
-    path.push({
-      state: stateMap.get(stateSerial)!,
-      event
-    });
-
-    const eventSerial = serializeEvent(event);
-    const { state: nextState, event: _nextEvent } = adjacency[
-      stateSerial
-    ].transitions[eventSerial];
-
-    if (!nextState) {
-      throw new Error(
-        `Invalid transition from ${stateSerial} with ${eventSerial}`
-      );
-    }
-    const prevState = stateMap.get(stateSerial);
-    const nextStateSerial = serializeState(
-      nextState,
-      event,
-      prevState
-    ) as SerializedState;
-    stateMap.set(nextStateSerial, nextState);
-
-    stateSerial = nextStateSerial;
-    state = nextState;
-  }
-
-  return {
-    state,
-    steps: path,
-    weight: path.length
+export interface AdjacencyValue<TState, TEvent> {
+  state: TState;
+  transitions: {
+    [key: SerializedEvent]: {
+      event: TEvent;
+      state: TState;
+    };
   };
 }
 
 export interface AdjacencyMap<TState, TEvent> {
-  [key: SerializedState]: {
-    state: TState;
-    transitions: {
-      [key: SerializedEvent]: {
-        event: TEvent;
-        state: TState;
-      };
-    };
-  };
+  [key: SerializedState]: AdjacencyValue<TState, TEvent>;
 }
 
 export function resolveTraversalOptions<TState, TEvent extends EventObject>(
